fine_tune_vllm/requirements.txt
streamlit
torch
torchvision
requests
Pillow